requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==4.4.2
executing==1.2.0
filelock==3.12.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.1
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
proglog==0.1.10
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
